StoryGen/requirements.txt
customtkinter~=5.1.2
Pillow~=9.5.0
requests~=2.28.2
openai~=0.27.4